Added introspection for default converted_status in convert_lead #105

+++ salesforce/tests/test_utils.py
@@ -2,7 +2,6 @@
 Tests for `salesforce.utils`
 """
 from django.test import TestCase
-from django.db import connections
 import unittest
 
 
@@ -26,23 +25,19 @@ class UtilitiesTest(TestCase):
         """
         lead = Lead(FirstName="Foo", LastName="Bar", Company="django-salesforce")
         lead.save()
-
-        # Get a status name setting for converted Leads
-        cur = connections['salesforce'].cursor()
-        cur.execute("SELECT MasterLabel FROM LeadStatus WHERE IsConverted=true")
-        converted_status = cur.fetchone()['MasterLabel']
-
-        ret = convert_lead(lead, converted_status=converted_status)
-        print("Response from convertLead: " +
-                ', '.join('%s: %s' % (k, v) for k, v in sorted(ret.items())))
-        expected_names = set(('accountId', 'contactId', 'leadId', 'opportunityId', 'success'))
-        self.assertEqual(set(ret), expected_names)
-        self.assertEqual(ret['success'], 'true')
-
-        # Cleaning up...
-        # Deleting the Account object will also delete the related Contact
-        # and Opportunity objects.
-        account = Account.objects.get(pk=ret['accountId'])
-        account.delete()
-
-        lead.delete()   # FYI, ret['leadId'] == lead.pk
+        ret = None
+        try:
+            ret = convert_lead(lead)
+            print("Response from convertLead: " +
+                    ', '.join('%s: %s' % (k, v) for k, v in sorted(ret.items())))
+            expected_names = set(('accountId', 'contactId', 'leadId', 'opportunityId', 'success'))
+            self.assertEqual(set(ret), expected_names)
+            self.assertEqual(ret['success'], 'true')
+        finally:
+            # Cleaning up...
+            if ret:
+                # Deleting the Account object will also delete the related Contact
+                # and Opportunity objects.
+                account = Account.objects.get(pk=ret['accountId'])
+                account.delete()
+            lead.delete()   # FYI, ret['leadId'] == lead.pk
